python/requirements.txt
# AI Trading Bot — Dependencies
# ================================

# Core
ccxt>=4.0.0
yfinance>=0.2.30
alpaca-trade-api>=3.0.0
oandapyV20>=0.7.2

# ML/AI
torch>=2.0.0
pytorch-forecasting>=1.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Technical Analysis
pandas-ta>=0.3.14
TA-Lib>=0.4.28

# NLP
finbert-embedding>=0.1.4
spacy>=3.7.0
nltk>=3.8.0
newspaper3k>=0.2.8
feedparser>=6.0.10

# Data
pandas>=2.1.0
numpy>=1.25.0
psycopg2-binary>=2.9.9
redis>=5.0.0
pyarrow>=14.0.0

# Monitoring & Dashboard
streamlit>=1.28.0
plotly>=5.18.0
python-telegram-bot>=20.6
structlog>=23.2.0

# Risk Analysis
empyrical>=0.5.5
scipy>=1.11.0

# Utils
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
praw>=7.7.0
requests>=2.31.0
apscheduler>=3.10.4

# ML Tracking
mlflow>=2.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.11.0
